--- analyzers/tests/SonarAnalyzer.Test/Constants.cs
@@ -1,32 +1,0 @@
-﻿/*
- * SonarAnalyzer for .NET
- * Copyright (C) 2015-2024 SonarSource SA
- * mailto: contact AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-
-namespace SonarAnalyzer.Test
-{
-    internal static class Constants
-    {
-        public const string NuGetLatestVersion = "LATEST";
-        public const string DotNet7Preview = "7.0.0-rc.1.22426.10";
-        public const string DotNet7PreviewAzureAppServices = "7.0.0-rc.1.22427.2";
-        public const string DotNetCore220Version = "2.2.0";
-        public const string WindowsLineEnding = "\r\n";
-        public const string UnixLineEnding = "\n";
-    }
-}
