+++ analyzers/tests/SonarAnalyzer.TestFramework/Common/AnalysisScaffolding.cs
