+++ analyzers/tests/SonarAnalyzer.TestFramework/Common/AssertIgnoreScope.cs
