+++ analyzers/tests/SonarAnalyzer.TestFramework/Common/FixAllDiagnosticProvider.cs
