+++ analyzers/tests/SonarAnalyzer.TestFramework/Common/Paths.cs
