+++ analyzers/tests/SonarAnalyzer.TestFramework/Common/TestContextHelper.cs
