+++ analyzers/tests/SonarAnalyzer.TestFramework/Compilation/ParseOptionsHelper.cs
