+++ analyzers/tests/SonarAnalyzer.TestFramework/Compilation/SnippetCompiler.cs
