+++ analyzers/tests/SonarAnalyzer.TestFramework/Extensions/CompilationExtensions.cs
