+++ analyzers/tests/SonarAnalyzer.TestFramework/Extensions/SyntaxTreeExtensions.cs
