+++ analyzers/tests/SonarAnalyzer.TestFramework/MetadataReferences/CoreMetadataReference.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-#if !NETFRAMEWORK
+#if NET
 
 using static SonarAnalyzer.Test.MetadataReferences.MetadataReferenceFactory;
