+++ analyzers/tests/SonarAnalyzer.TestFramework/MetadataReferences/FrameworkMetadataReference.cs
