+++ analyzers/tests/SonarAnalyzer.TestFramework/MetadataReferences/NuGetMetadataFactory.cs
