+++ analyzers/tests/SonarAnalyzer.TestFramework/MetadataReferences/WindowsFormsMetadataReference.cs
