+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/CodeFixVerifier.cs
