+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/DiagnosticVerifier.cs
