+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/DiagnosticsComparer.cs
