+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/IIssueLocation.cs
