+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/IssueLocationCollector.cs
