+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/UnexpectedDiagnosticException.cs
