+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/Verifier.cs
