+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/VerifierBuilder.cs
